Corrige bug #1

+++ 05-spring-ai/src/main/java/dio/budgeting/domain/Transaction.java
@@ -1,5 +1,7 @@
 package dio.budgeting.domain;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -8,10 +10,10 @@ import lombok.Getter;
 public class Transaction {
     private TransactionId id;
     private String description;
-    private long amount;
+    private BigDecimal amount;
     private Category category;
 
-    public Transaction(String description, long amount, Category category) {
+    public Transaction(String description, BigDecimal amount, Category category) {
         this.id = new TransactionId();
         this.description = description;
         this.amount = amount;
